fix(c1): relocate testcontainer-using resolver tests to integration
C1's persistence tests (cache-hit, insert-and-return, multi-provider,
no-match, partial-match, concurrent-race, timeout) all use the real Postgres
session fixture and exercise INSERT ... ON CONFLICT semantics. Those are
integration concerns. Pure-function _normalize stays under tests/unit/. Drops
tests/unit/conftest.py since tests/integration/conftest.py already provides
db_session for the relocated tests.
